requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
reportlab==4.0.7
Pillow==10.1.0
pyinstaller==6.3.0
